Rewrite file creator.
