+++ src/extension.ts
@@ -4,10 +4,9 @@ import * as fs from "fs";
 import * as path from "path";
 import * as vscode from "vscode";
 
-import * as constants from "./constants";
-import { CancelError } from "./customErrors";
-import { FileCreator, IFileCreatorInputData } from "./fileCreator";
+import { createFiles } from "./fileCreator";
 import { getUserInput } from "./inputs";
+import { handleError } from './errorHandler';
 
 export function activate(context: vscode.ExtensionContext) {
 
@@ -25,29 +24,12 @@ export function activate(context: vscode.ExtensionContext) {
 
     try {
       const userInput = await getUserInput(templateFolderRawPaths);
-
-      const data: IFileCreatorInputData = {
-        templateFolderPath: userInput.selectedTemplatePath,
-        pathToCreateAt: directoryPath,
-        inputName: userInput.inputName
-      }
-
-      const fileCreator = new FileCreator(data);
-      await fileCreator.createFiles();
-
+      await createFiles(userInput, directoryPath);
     } catch (error) {
-      if (error instanceof CancelError) { return; }
-
-      const message: string = error.message ? error.message : "There was a problem creating your file(s).";
-      const isModal = message.startsWith(constants.ERROR_SETUP_MESSAGE_PREFIX);
-
-      const errorMessage = error.message ? error.message : "There was a problem creating your file(s).";
-      vscode.window.showErrorMessage(errorMessage, { modal: isModal });
+      handleError(error);
     }
 
   });
 
   context.subscriptions.push(disposable);
 }
-
-// export function deactivate() {}
